requirements.txt
fastapi
uvicorn
pdfplumber
nltk
spacy
chromadb
sentence-transformers
google-generativeai
python-dotenv